rearrange library structure
